+++ crates/configuration/src/version5/mod.rs
@@ -1,18 +1,18 @@
 //! Internal Configuration and state for our connector.
 
-pub mod comparison;
+mod comparison;
 pub mod connection_settings;
 pub mod metadata;
 pub mod native_operations;
-pub mod options;
+mod options;
 mod to_runtime_configuration;
-mod upgrade_from_v3;
+mod upgrade_from_v4;
 
 use std::borrow::Cow;
 use std::collections::HashSet;
 use std::path::Path;
 pub use to_runtime_configuration::make_runtime_configuration;
-pub use upgrade_from_v3::upgrade_from_v3;
+pub use upgrade_from_v4::upgrade_from_v4;
 
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
@@ -57,7 +57,7 @@ pub struct ParsedConfiguration {
 
 #[derive(Clone, PartialEq, Eq, Debug, Deserialize, Serialize, JsonSchema)]
 pub enum Version {
-    #[serde(rename = "4")]
+    #[serde(rename = "5")]
     This,
 }
 
@@ -99,16 +99,18 @@ fn get_type_ndc_name(r#type: &metadata::Type) -> &str {
 /// In order to ensure that our schema ends up including exactly the relevant types we need to
 /// inform the introspection query of the types that may appear in native queries, such that they
 /// may be included.
-fn native_query_field_types(native_queries: &metadata::NativeQueries) -> Vec<String> {
+fn native_operations_field_types(native_operations: &metadata::NativeOperations) -> Vec<String> {
     let mut result: HashSet<&str> = HashSet::new();
 
-    for native_query in native_queries.0.values() {
-        for argument in native_query.arguments.values() {
-            result.insert(get_type_ndc_name(&argument.r#type));
-        }
+    for operations in [&native_operations.queries, &native_operations.mutations] {
+        for native_query in operations.0.values() {
+            for argument in native_query.arguments.values() {
+                result.insert(get_type_ndc_name(&argument.r#type));
+            }
 
-        for column in native_query.columns.values() {
-            result.insert(get_type_ndc_name(&column.r#type));
+            for column in native_query.columns.values() {
+                result.insert(get_type_ndc_name(&column.r#type));
+            }
         }
     }
 
@@ -153,7 +155,9 @@ pub async fn introspect(
         .bind(serde_json::to_value(
             &args.introspection_options.type_representations,
         )?)
-        .bind(native_query_field_types(&args.metadata.native_queries));
+        .bind(native_operations_field_types(
+            &args.metadata.native_operations,
+        ));
 
     let row = connection
         .fetch_one(query)
@@ -179,9 +183,11 @@ pub async fn introspect(
         connection_settings: args.connection_settings,
         metadata: metadata::Metadata {
             tables,
-            scalar_types,
-            composite_types,
-            native_queries: args.metadata.native_queries,
+            types: metadata::Types {
+                scalar: scalar_types,
+                composite: composite_types,
+            },
+            native_operations: args.metadata.native_operations,
         },
         introspection_options: args.introspection_options,
         mutations_version: args.mutations_version,
@@ -213,13 +219,18 @@ pub async fn parse_configuration(
             message: error.to_string(),
         })?;
     // look for native query sql file references and read from disk.
-    for native_query_sql in parsed_config.metadata.native_queries.0.values_mut() {
-        native_query_sql.sql = metadata::NativeQuerySqlEither::NativeQuerySql(
-            native_query_sql
-                .sql
-                .from_external(configuration_dir.as_ref())
-                .map_err(ParseConfigurationError::IoErrorButStringified)?,
-        );
+    for operations in [
+        &mut parsed_config.metadata.native_operations.queries,
+        &mut parsed_config.metadata.native_operations.mutations,
+    ] {
+        for native_query_sql in operations.0.values_mut() {
+            native_query_sql.sql = metadata::NativeQuerySqlEither::NativeQuerySql(
+                native_query_sql
+                    .sql
+                    .from_external(configuration_dir.as_ref())
+                    .map_err(ParseConfigurationError::IoErrorButStringified)?,
+            );
+        }
     }
 
     Ok(parsed_config)
@@ -243,26 +254,31 @@ pub async fn write_parsed_configuration(
     .await?;
 
     // look for native query sql file references and write them to disk.
-    for native_query_sql in parsed_config.metadata.native_queries.0.values() {
-        if let metadata::NativeQuerySqlEither::NativeQuerySql(
-            metadata::NativeQuerySql::FromFile { file, sql },
-        ) = &native_query_sql.sql
-        {
-            if file.is_absolute() || file.starts_with("..") {
-                Err(
-                    WriteParsedConfigurationError::WritingOutsideDestinationDir {
-                        dir: out_dir.as_ref().to_owned(),
-                        file: file.clone(),
-                    },
-                )?;
-            };
+    for operations in [
+        &parsed_config.metadata.native_operations.queries,
+        &parsed_config.metadata.native_operations.mutations,
+    ] {
+        for native_query_sql in operations.0.values() {
+            if let metadata::NativeQuerySqlEither::NativeQuerySql(
+                metadata::NativeQuerySql::FromFile { file, sql },
+            ) = &native_query_sql.sql
+            {
+                if file.is_absolute() || file.starts_with("..") {
+                    Err(
+                        WriteParsedConfigurationError::WritingOutsideDestinationDir {
+                            dir: out_dir.as_ref().to_owned(),
+                            file: file.clone(),
+                        },
+                    )?;
+                };
 
-            let native_query_file = out_dir.as_ref().to_owned().join(file);
-            if let Some(native_query_sql_dir) = native_query_file.parent() {
-                fs::create_dir_all(native_query_sql_dir).await?;
+                let native_query_file = out_dir.as_ref().to_owned().join(file);
+                if let Some(native_query_sql_dir) = native_query_file.parent() {
+                    fs::create_dir_all(native_query_sql_dir).await?;
+                };
+                fs::write(native_query_file, String::from(sql.clone())).await?;
             };
-            fs::write(native_query_file, String::from(sql.clone())).await?;
-        };
+        }
     }
 
     // create the jsonschema file